requirements.txt
# Requirements for Claude Code Agentic System
Flask==2.3.2
Flask-SocketIO==5.3.0
python-socketio==5.8.0
PyYAML==6.0.1
requests==2.31.0

# Optional dependencies (install if available)
# claude-code  # Install separately from Anthropic